+++ CHANGELOG.md
@@ -35,6 +35,8 @@
   - [#4428](https://github.com/wix/ricos/pull/4428) remove fallback config from FloatingAddPluginMenu
 - `modal-service`
   - [#4427](https://github.com/wix/ricos/pull/4427) modals re-renders on content changes
+- `gallery`
+  - [#4419](https://github.com/wix/ricos/pull/4419) add modal button option
 
 ## 8.72.31 (Aug 21, 2022)